# api/models/Customer.js
class Customer {
    constructor(customer) {
        this.name = customer.name;
        this.email = customer.email;
        this.address = customer.address;
        this.phoneNumber = customer.phoneNumber;
    }
}

module.exports = Customer;